Add MTD configuration for leader network data.
