+++ etc/visual-studio/libopenthread_k.vcxproj.filters
@@ -156,6 +156,9 @@
     <ClCompile Include="..\..\src\core\thread\network_data_leader.cpp">
       <Filter>Source Files\thread</Filter>
     </ClCompile>
+    <ClCompile Include="..\..\src\core\thread\network_data_leader_ftd.cpp">
+      <Filter>Source Files\thread</Filter>
+    </ClCompile>
     <ClCompile Include="..\..\src\core\thread\network_data_local.cpp">
       <Filter>Source Files\thread</Filter>
     </ClCompile>
